Testing defects fix (#53)
* footer fix for register page

* removes profile page from user menu

* submit on enter keybutton pressed

* list "description" fix

* Styling update

Updates the styling around spotlight courses and more like this course bars. Sets the footer to be absolutely placed. Default screen size is the height of the screen.

* Updates Menu

Updates the styling for the signed in menu button to categorize groupings of buttons / links.

Updates test for Header.js and UserMenu.js

* login error

* Removed the use of useLayoutEffect

* Updates import

* refactors login page

* Update styling

* rearrange userMenu order

* updates code styling

* Updates Eslint

* Updates component tests for user menu

Co-authored-by: rglasgow-deloitte <rglasgow@deloitte.com>